issue#2065. Remarks correction
Creating Directory instead of Switch. Creating JsonModel with
initialization
